Enforcing payment driver methods using an interface
